chore(tests): migrate jsii tests to jest
In an effort to try and stop using `nodeunit` everywhere, migrating all
existing tests in the `jsii` module to `jest`.
