+++ packages/jsii/test/enums.test.ts
@@ -1,0 +1,39 @@
+import { sourceToAssemblyHelper } from '../lib';
+
+// ----------------------------------------------------------------------
+test('test parsing enum with two members and no values', async () => {
+  const assembly = await sourceToAssemblyHelper(`
+    export enum Foo {
+      BAR,
+      BAZ
+    }
+  `);
+
+  expect(assembly.types!['testpkg.Foo']).toEqual({
+    assembly: 'testpkg',
+    fqn: 'testpkg.Foo',
+    kind: 'enum',
+    members: [{ name: 'BAR' }, { name: 'BAZ' }],
+    locationInModule: { filename: 'index.ts', line: 2 },
+    name: 'Foo'
+  });
+});
+
+// ----------------------------------------------------------------------
+test('test parsing enum with two members and assigned values', async () => {
+  const assembly = await sourceToAssemblyHelper(`
+    export enum Foo {
+      BAR = 'Bar',
+      BAZ = 'Baz'
+    }
+  `);
+
+  expect(assembly.types!['testpkg.Foo']).toEqual({
+    assembly: 'testpkg',
+    fqn: 'testpkg.Foo',
+    kind: 'enum',
+    members: [{ name: 'BAR' }, { name: 'BAZ' }],
+    locationInModule: { filename: 'index.ts', line: 2 },
+    name: 'Foo'
+  });
+});
